requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-form==1.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.1.3
WTForms==3.2.1